requirements.txt
appdirs==1.4.4
Appium-Python-Client==0.52
attrs==19.3.0
black==19.10b0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
coloredlogs==14.0
executor==23.1
fasteners==0.15
humanfriendly==8.2
idna==2.9
monotonic==1.5
numpy==1.18.4
opencv-python==4.2.0.34
pathspec==0.8.0
property-manager==3.0
pyreadline==2.1
regex==2020.5.14
selenium==3.141.0
six==1.14.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
verboselogs==1.7
